Congressional App Challenge Flask Code/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
tensorflow-cpu==2.9.0
pandas==1.4.2
Werkzeug==2.2.2
pillow==9.1.0
scipy==1.8.0
